Modify UI and reverse TransactionListPanel
Previously, each transaction would be appended to the bottom of the
transaction list panel, which would impede the user from getting
immediate response as to their input, as the TransactionListPanel does
not scroll automatically.

Rather than making the panel auto-scroll, let's reverse
the order of the list so that the user is always looking at the most
recent transaction.

Additionally, let's modify the CommandBox to span the whole Window
with a smaller font size so that it can accomodate a longer input
command.