Add updateExpenseSummary to Model and ModelManager
In order to interact with the UI, ModelManager needs an attribute
to allow the UI to directly get the required data for plotting the
pie chart.

Let's add an `expenseSummary` HashMap<String, Double> for this. The
`expenseSummary` hashmap maps the expense Category to a value, which
represents the total expense for that category. The
updateExpenseSummary method will update the `expenseSummary` hashmap
when there are changes in the transactions. This warrants the need
for other methods that directly modify the transactions in UniCash
to also be updated to call this method when they make changes in the
transactions. Unit tests that test the added methods have also been
added.

Note 1: `UniCash#getSumOfExpensePerCategory` is the method that
calculates the expenses per category, while
`Model#updateExpenseSummary` simply uses the result of the
calculations to update `expenseSummary`.
Note 2: The `AddTransactionCommandTest` has been modified because
the ModelStub class has to implement the
`Model#updateExpenseSummary` prototype.