Link `expenseSummary` to `SummaryWindow`
`SummaryWindow` is currently plotting fictional (hardcoded) pie
chart data. Thus, it is not plotting the true data in UniCash.

Let's link the `expenseSummary` attribute of UniCash to
`SummaryWindow`. Since `MainWindow` is the UiPart that links `Logic`
and `SummaryWindow` together, we first modify the `SummaryWindow`
constructor to take in a `HashMap<String, Double> expenseSummary`.
Then, within `MainWindow`, when constructing `SummaryWindow`, we
pass in the `expenseSummary` hash map as an input to link
`MainWindow` and `SummaryWindow`. Next, we link `expenseSummary` to
`MainWindow` by creating the `getExpenseSummary` method in `Logic`
and `LogicManager`. Lastly, we edit `SummaryWindow#setPieChart` to
modify the pie chart's data using `expenseSummary`, instead of the
fictional (hardcoded) data.

Note 1: `MainWindow#handleSummary` now calls
`SummaryWindow#setPieChart` so that when the user executes the
"summary" command, the pie chart will be updated accordingly.
Note 2: The current implementation uses for-loops extensively. This
will later be modified to use Java streams instead.