Modify `MainWindow` and `SummaryWindow` for better abstraction
`SummaryWindow#setPieChart` now does not take in any parameters. It
is able to directly access `UniCash#expenseSummary`. It may be
better to restrict `SummaryWindow`'s access to `expenseSummary`.

Let's modify `SummaryWindow#setPieChart` to take in `expenseSummary`
as a parameter instead. `MainWindow`'s initialization of
`SummaryWindow` also has to be updated due to `SummaryWindow`'s
constructor changing.