Modify test cases due to changes to Category constructor
There are several test cases for LogicManager, ModelManager, and UniCash
that are failing due to the changes to the Category(String) constructor.
In particular, categories have been modified by Rubesh to be set to
lowercase. This causes the expected behavior of test cases that rely on
this constructor to change.

Let's modify the test cases to reflect this change in expected behavior.
We simply modify the expected categories to have all lowercase chars.