Add logic to calculate expense per year-month to Model component
The current implementation does not support calculating the summary
statistics for the year-month level. This is needed for creating the
line chart in the summary window.

Let's add the `UniCash#getSumOfExpensePerYearMonth` to calculate the
total expense per year-month. The logic to filter away all income-type
transactions has also been extracted away to the `getAllExpense` method.
To make it more convenient to get the transaction's year-month value,
let's also make slight updates to the `DateTime` class from the `Model`
component. In particular, let's create a `DateTime#getYearMonth` method
that will return a `YearMonth` object.

The modifications made to `UniCash` have been made to shorten the code
for `UniCash#getSumOfExpensePerCategory`, which improves readability.
Furthermore, the logic to handle transactions with and without
categories has also been extracted to private methods also to improve
readability. Namely, these methods are `handleNoCategoryTransaction` and
`handleCategoryTransaction`.