Add functionality to pass expense summary per year-month to `MainWindow`
The expense summary per year-month is calculated by `UniCash` but is not
integrated with the rest of the other components.

Let's modify the `Logic` and `Model` components to pass the expense
summary from `UniCash` to `MainWindow` for displaying on the
`SummaryWindow`'s pop-up window.

Note: The design for the line graph in `SummaryWindow` has not been
completed as of this commit.