Update `SummaryWindow` to add line chart and new icon for window
The summary window now supports only the pie chart. It also uses the
help window's icon.

Let's integrate the line chart into the summary window, and add a new
icon for the window instead of using the help window's icon.

The window's icon should be changed because the help window's icon
contains a '?' symbol, which might confuse users.