Disable the summary window from opening if there are no expenses
The summary window opens every time the user makes the "summary"
command. However, if there are no expenses to show, then there is no
reason to open the summary window.

Let's prevent the summary window from opening if there are no expenses.
However, if the summary window was already opened before, then this fix
will not close the summary window. It only applies if the summary window
was not opened before the "summary" command was made.

This was intentionally designed to behave this way because we do not
want to close the summary window when all expenses have been removed. If
we choose to do so, then it might confuse the user as the window might
suddenly disappear.