Fix bug where `summary` output message is wrong when there is income
There is a bug whereby the output message of the "summary" command
is "Opened UniCa$h summary window." even if there are no expenses,
but there are income transactions. This is caused by the
`SummaryCommand#execute` method implementation, which only checks if
there are transactions available, not if there are expense
transactions available.

Let's fix this bug by changing the check in `SummaryCommand#execute`
and instead, check if there are expense transactions available. To
do this, we create a new method in the Model class called
`hasExpenses`. This calls the new `UniCash#hasExpenses` method. We
also implement the `Model#hasExpenses` function prototype in the
other classes that implement the `Model` interface. Lastly, let's
add additional tests to bring maintain the current level of code
coverage.

Creating this new function prototype in `Model` allows minimal
modifications to the `SummaryCommand` class, as its `execute` method
only needs to call the `Model#hasExpenses` method. This maintains
readability of the `SummaryCommand#execute` method.